Split realtime handlers by backend
